refactor : N+1 개선: Comment 카운트를 countByCommunityIds 배치 집계로 전환해 목록 조회에서 카운트 쿼리 1회로 축소 + 타입 불일치 완벽 해결: user_history.history_id + 빌드테스트

+++ src/main/java/tave/crezipsa/crezipsa/application/trend/dto/response/UserHistoryResponse.java
@@ -3,8 +3,8 @@ package tave.crezipsa.crezipsa.application.trend.dto.response;
 import tave.crezipsa.crezipsa.domain.user.entity.UserHistory;
 
 public record UserHistoryResponse (
-        long historyId,
-        String history
+		int historyId,
+		String history
 ){
     public static UserHistoryResponse from(UserHistory h) {
         return new UserHistoryResponse(h.getHistoryId(), h.getHistoryContent());
